Added JOML and did some changes to the model implementation.

+++ client/src/main/java/dev/zyko/starfight/client/gui/GuiComponent.java
@@ -2,6 +2,6 @@ package dev.zyko.starfight.client.gui;
 
 public interface GuiComponent {
 
-    void drawComponent(double mouseX, double mouseY, float partialTicks);
+    void drawComponent(double mouseX, double mouseY, double partialTicks);
 
 }
